shrek is viable
Shrek is a melee-strength hero specialising in lengthy single target
disables and ganking prowess. Shrek is most feared for his devastating
ultimate "Shrek is love" and its sub-ability "Shrek is life".
Q: Spread Buttcheeks
Shrek teleports behind an enemy unit, grabs their asshole and spreads
their cheeks, increasing the damage dealt by all of Shrek's abilities by
15/25/35/45% and stunning them for 1 second. This ability has no
cooldown. Manacost: 80/90/100/110. Range 500
W: This is my Swamp
Shrek passively causes damage to all who look at him in a 900 radius. If
an enemy is facing within a 45 degree angle of him while having vision
they take 5/10/15/20 damage per second and are silenced if he comes
within 200 units of them. Shrek can activate this ability to gain flying
movement and vision for 30 seconds. Cooldown 60 seconds. Manacost: 200
E: It's all Ogre Now
Shrek vanishes after every victim is used. Gaining maximum movement
speed, phase, terrain ignoring invisibility for 4/5/6/7 every time he
kills an enemy hero with Shrek is Love. If he kills an enemy with Shrek
is Love, then all ability cooldowns are reduced by 25/50/75/100% as his
lust has not been satisfied.
R: Shrek is love
Shrek grabs hold of an enemy hero, stunning them for 3/4/5 seconds and
has his way with them, dealing 600 physical damage per second to them
for 2/3/4 seconds. If an enemy is not killed by the effects of Shrek is
love, then the effects of Shrek is life are applied to that target. If
Shrek does not kill an enemy with this ability, then his abilities are
all muted until Shrek is Love is off cooldown. Cooldown: 90/80/70
Manacost: 300/400/500
D: Shrek is life
Shrek has satisfied his lust in a target and they are stained with the
essence of Shrek and this culminates in them. A small Shrek grows within
the target and after 9 minutes of Shrek is Life affecting the target
they are stunned for 20/30/40 seconds and give birth to the little Shrek
after 15/25/35 seconds of being stunned. The enemy is visable while
stunned and the duration of the debuff is not visable to the enemy.
The small Shrek can use all of Shrek's abilities but only deal 50% of
the damage and cannot apply Shrek is Life's affects to any enemy hero.
The small Shrek shares all of Shrek's stats at a 50% rate rounded down.